+++ tests/Framework/Http/RequestFactoryTest.php
@@ -11,14 +11,14 @@ declare(strict_types=1);
 
 namespace Spiral\Tests\Framework\Http;
 
+use Nyholm\Psr7\Factory\Psr17Factory;
 use PHPUnit\Framework\TestCase;
-use Spiral\Http\Diactoros\ServerRequestFactory;
 
 class RequestFactoryTest extends TestCase
 {
     public function testCreateRequest(): void
     {
-        $r = (new ServerRequestFactory())->createServerRequest('GET', '/home');
+        $r = (new Psr17Factory())->createServerRequest('GET', '/home');
         $this->assertSame('GET', $r->getMethod());
         $this->assertSame('/home', $r->getUri()->getPath());
     }